+++ R/create_trackhub_description.R
@@ -1,0 +1,12 @@
+create_trackhub_description <- function(file,
+                                        trackdbFile,
+                                        verbose=TRUE){
+    
+    descriptionUrl <-  file.path(dirname(trackdbFile),
+                                 paste0(basename(file),".html"))
+    messager("Creating file:",basename(descriptionUrl),v=verbose) 
+    
+    writeLines(text = basename(file), 
+               con = descriptionUrl)
+    return(descriptionUrl)
+}