+++ R/list_to_char.R
@@ -1,0 +1,7 @@
+list_to_char <- function(lst,
+                         sep="\t",
+                         newline="\n"){ 
+    paste(names(lst), lst, 
+          sep = sep, 
+          collapse = newline)   
+}